+++ packages/webview_flutter/webview_flutter_android/android/src/main/java/io/flutter/plugins/webviewflutter/WebViewFlutterPlugin.java
@@ -10,6 +10,25 @@ import io.flutter.embedding.engine.plugins.FlutterPlugin;
 import io.flutter.embedding.engine.plugins.activity.ActivityAware;
 import io.flutter.embedding.engine.plugins.activity.ActivityPluginBinding;
 
+import io.flutter.plugin.common.BinaryMessenger;
+import io.flutter.plugin.platform.PlatformViewRegistry;
+import io.flutter.plugins.webviewflutter.GeneratedAndroidWebView.CookieManagerHostApi;
+import io.flutter.plugins.webviewflutter.GeneratedAndroidWebView.CustomViewCallbackHostApi;
+import io.flutter.plugins.webviewflutter.GeneratedAndroidWebView.DownloadListenerHostApi;
+import io.flutter.plugins.webviewflutter.GeneratedAndroidWebView.FlutterAssetManagerHostApi;
+import io.flutter.plugins.webviewflutter.GeneratedAndroidWebView.GeolocationPermissionsCallbackHostApi;
+import io.flutter.plugins.webviewflutter.GeneratedAndroidWebView.HttpAuthHandlerHostApi;
+import io.flutter.plugins.webviewflutter.GeneratedAndroidWebView.InstanceManagerHostApi;
+import io.flutter.plugins.webviewflutter.GeneratedAndroidWebView.JavaObjectHostApi;
+import io.flutter.plugins.webviewflutter.GeneratedAndroidWebView.JavaScriptChannelHostApi;
+import io.flutter.plugins.webviewflutter.GeneratedAndroidWebView.PermissionRequestHostApi;
+import io.flutter.plugins.webviewflutter.GeneratedAndroidWebView.SslErrorHandlerHostApi;
+import io.flutter.plugins.webviewflutter.GeneratedAndroidWebView.WebChromeClientHostApi;
+import io.flutter.plugins.webviewflutter.GeneratedAndroidWebView.WebSettingsHostApi;
+import io.flutter.plugins.webviewflutter.GeneratedAndroidWebView.WebStorageHostApi;
+import io.flutter.plugins.webviewflutter.GeneratedAndroidWebView.WebViewClientHostApi;
+import io.flutter.plugins.webviewflutter.GeneratedAndroidWebView.WebViewHostApi;
+
 /**
  * Java platform implementation of the webview_flutter plugin.
  *
@@ -46,6 +65,63 @@ public class WebViewFlutterPlugin implements FlutterPlugin, ActivityAware {
             new FlutterViewFactory(proxyApiRegistrar.getInstanceManager()));
 
     proxyApiRegistrar.setUp();
+    
+    webViewHostApi =
+        new WebViewHostApiImpl(
+            instanceManager, binaryMessenger, new WebViewHostApiImpl.WebViewProxy(), context);
+    javaScriptChannelHostApi =
+        new JavaScriptChannelHostApiImpl(
+            instanceManager,
+            new JavaScriptChannelHostApiImpl.JavaScriptChannelCreator(),
+            new JavaScriptChannelFlutterApiImpl(binaryMessenger, instanceManager),
+            new Handler(context.getMainLooper()));
+
+    JavaObjectHostApi.setup(binaryMessenger, new JavaObjectHostApiImpl(instanceManager));
+    WebViewHostApi.setup(binaryMessenger, webViewHostApi);
+    JavaScriptChannelHostApi.setup(binaryMessenger, javaScriptChannelHostApi);
+    WebViewClientHostApi.setup(
+        binaryMessenger,
+        new WebViewClientHostApiImpl(
+            instanceManager,
+            new WebViewClientHostApiImpl.WebViewClientCreator(),
+            new WebViewClientFlutterApiImpl(binaryMessenger, instanceManager)));
+    WebChromeClientHostApi.setup(
+        binaryMessenger,
+        new WebChromeClientHostApiImpl(
+            instanceManager,
+            new WebChromeClientHostApiImpl.WebChromeClientCreator(),
+            new WebChromeClientFlutterApiImpl(binaryMessenger, instanceManager)));
+    DownloadListenerHostApi.setup(
+        binaryMessenger,
+        new DownloadListenerHostApiImpl(
+            instanceManager,
+            new DownloadListenerHostApiImpl.DownloadListenerCreator(),
+            new DownloadListenerFlutterApiImpl(binaryMessenger, instanceManager)));
+    WebSettingsHostApi.setup(
+        binaryMessenger,
+        new WebSettingsHostApiImpl(
+            instanceManager, new WebSettingsHostApiImpl.WebSettingsCreator()));
+    FlutterAssetManagerHostApi.setup(
+        binaryMessenger, new FlutterAssetManagerHostApiImpl(flutterAssetManager));
+    CookieManagerHostApi.setup(
+        binaryMessenger, new CookieManagerHostApiImpl(binaryMessenger, instanceManager));
+    WebStorageHostApi.setup(
+        binaryMessenger,
+        new WebStorageHostApiImpl(instanceManager, new WebStorageHostApiImpl.WebStorageCreator()));
+
+    if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.LOLLIPOP) {
+      PermissionRequestHostApi.setup(
+          binaryMessenger, new PermissionRequestHostApiImpl(binaryMessenger, instanceManager));
+    }
+    GeolocationPermissionsCallbackHostApi.setup(
+        binaryMessenger,
+        new GeolocationPermissionsCallbackHostApiImpl(binaryMessenger, instanceManager));
+    CustomViewCallbackHostApi.setup(
+        binaryMessenger, new CustomViewCallbackHostApiImpl(binaryMessenger, instanceManager));
+    HttpAuthHandlerHostApi.setup(
+        binaryMessenger, new HttpAuthHandlerHostApiImpl(binaryMessenger, instanceManager));
+    SslErrorHandlerHostApi.setup(
+        binaryMessenger, new SslErrorHandlerHostApiImpl(binaryMessenger, instanceManager));
   }
 
   @Override