+++ packages/webview_flutter/webview_flutter_android/test/android_navigation_delegate_test.mocks.dart
@@ -182,3 +182,47 @@ class MockDownloadListener extends _i1.Mock implements _i2.DownloadListener {
         ),
       ) as _i2.DownloadListener);
 }
+
+/// A class which mocks [SslErrorHandler].
+///
+/// See the documentation for Mockito's code generation for more information.
+class MockSslErrorHandler extends _i1.Mock implements _i2.SslErrorHandler {
+  MockSslErrorHandler() {
+    _i1.throwOnMissingStub(this);
+  }
+
+  @override
+  _i4.Future<void> cancel() => (super.noSuchMethod(
+        Invocation.method(
+          #cancel,
+          [],
+        ),
+        returnValue: _i4.Future<void>.value(),
+        returnValueForMissingStub: _i4.Future<void>.value(),
+      ) as _i4.Future<void>);
+
+  @override
+  _i4.Future<void> proceed() => (super.noSuchMethod(
+        Invocation.method(
+          #proceed,
+          [],
+        ),
+        returnValue: _i4.Future<void>.value(),
+        returnValueForMissingStub: _i4.Future<void>.value(),
+      ) as _i4.Future<void>);
+
+  @override
+  _i2.JavaObject copy() => (super.noSuchMethod(
+        Invocation.method(
+          #copy,
+          [],
+        ),
+        returnValue: _FakeJavaObject_0(
+          this,
+          Invocation.method(
+            #copy,
+            [],
+          ),
+        ),
+      ) as _i2.JavaObject);
+}