+++ packages/webview_flutter/webview_flutter_wkwebview/lib/src/webkit_webview_controller.dart
@@ -867,6 +867,13 @@ window.addEventListener("error", function(e) {
           onJavaScriptTextInputDialog) async {
     _onJavaScriptTextInputDialog = onJavaScriptTextInputDialog;
   }
+
+  /// Sets the trusted custom url schemes for the web view.
+  Future<void> setCustomUrlSchemes(Set<String> customUrlSchemes) async {
+    return _webView.configuration.setCustomUrlSchemes(<String>{
+      for (final String urlScheme in customUrlSchemes) urlScheme.toLowerCase()
+    });
+  }
 }
 
 /// An implementation of [JavaScriptChannelParams] with the WebKit api.
@@ -1170,7 +1177,7 @@ class WebKitNavigationDelegate extends PlatformNavigationDelegate {
           NSUrlSessionAuthChallengeDisposition disposition,
           NSUrlCredential? credential,
         ) completionHandler,
-      ) {
+      ) async {
         if (challenge.protectionSpace.authenticationMethod ==
                 NSUrlAuthenticationMethod.httpBasic ||
             challenge.protectionSpace.authenticationMethod ==
@@ -1206,6 +1213,42 @@ class WebKitNavigationDelegate extends PlatformNavigationDelegate {
             );
             return;
           }
+        } else if (challenge.protectionSpace.authenticationMethod ==
+            NSUrlAuthenticationMethod.serverTrust) {
+          final FutureOr<SslErrorDecision> Function(SslError)? callback =
+              weakThis.target?._onSslError;
+          if (callback != null) {
+            final sslErrorDecision = await callback(
+              SslError(
+                host: challenge.protectionSpace.host!,
+                scheme: challenge.protectionSpace.protocol!,
+                port: challenge.protectionSpace.port!,
+                errorType: challenge.protectionSpace.sslErrorType,
+                certificate: SslCertificate(
+                  issuedBy: null,
+                  issuedTo: null,
+                  validNotAfterDate: null,
+                  validNotBeforeDate: null,
+                  x509CertificateDer:
+                      challenge.protectionSpace.x509CertificateDer,
+                ),
+              ),
+            );
+
+            if (sslErrorDecision == SslErrorDecision.proceed) {
+              completionHandler(
+                NSUrlSessionAuthChallengeDisposition.useCredential,
+                null,
+              );
+            } else if (sslErrorDecision == SslErrorDecision.cancel) {
+              completionHandler(
+                NSUrlSessionAuthChallengeDisposition
+                    .cancelAuthenticationChallenge,
+                null,
+              );
+            }
+            return;
+          }
         }
 
         completionHandler(
@@ -1227,6 +1270,7 @@ class WebKitNavigationDelegate extends PlatformNavigationDelegate {
   NavigationRequestCallback? _onNavigationRequest;
   UrlChangeCallback? _onUrlChange;
   HttpAuthRequestCallback? _onHttpAuthRequest;
+  SslErrorCallback? _onSslError;
 
   @override
   Future<void> setOnPageFinished(PageEventCallback onPageFinished) async {
@@ -1272,6 +1316,11 @@ class WebKitNavigationDelegate extends PlatformNavigationDelegate {
     HttpAuthRequestCallback onHttpAuthRequest,
   ) async {
     _onHttpAuthRequest = onHttpAuthRequest;
+  }
+
+  @override
+  Future<void> setOnSslError(SslErrorCallback onSslError) async {
+    _onSslError = onSslError;
   }
 }