+++ packages/webview_flutter/webview_flutter_platform_interface/CHANGELOG.md
@@ -2,6 +2,11 @@
 
 * Updates minimum supported SDK version to Flutter 3.22/Dart 3.4.
 
+## 2.11.0
+
+* Adds support for handling SSL errors. See `PlatformNavigationDelegate.setOnSslError`.
+* Updates minimum supported SDK version to Flutter 3.19/Dart 3.3.
+
 ## 2.10.0
 
 * Adds `WebResourceRequest` and `WebResourceResponse` to `HttpResponseError`.